fix(Vercel): separate the authOptions into another file
