+++ src/app/api/whoAmI/route.ts
@@ -1,7 +1,7 @@
 import { getServerSession } from "next-auth";
 import { NextResponse } from "next/server";
 
-import { authOptions } from "../auth/[...nextauth]/route";
+import { authOptions } from "@/config/authOptions";
 
 export async function GET() {
   const session = await getServerSession(authOptions);
